+++ CHANGELOG.md
@@ -7,3 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- The experiment builder now can handle facts-total (categorized as a 'framework module' for now) and extremesealevel modules (currently just extremesealevel-pointsoverthreshold) ([PR 10](https://github.com/fact-sealevel/facts-experiment-builder/pull/10), [@e-marshall](https://github.com/e-marshall))
+
+